Move rust-install tests to tests directory
